+++ src/core/interpreter/index.ts
@@ -36,7 +36,7 @@ import {
     BrsInterface,
     toAssociativeArray,
 } from "../brsTypes";
-import { tryCoerce } from "../brsTypes/coercion";
+import { tryCoerce } from "../brsTypes/Coercions";
 import { shared, stats } from "..";
 import { Lexeme, GlobalFunctions } from "../lexer";
 import { isToken, Location } from "../lexer/Token";
